@@ -1,59 +1,60 @@
 {
-  "name": "node-lame",
-  "description": "LAME MP3 encoder for Node.js",
-  "version": "1.1.0",
-  "homepage": "https://github.com/jankarres/node-lame",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jankarres/node-lame.git"
-  },
-  "bugs": {
-    "url": "https://github.com/jankarres/node-lame/issues"
-  },
-  "author": {
-    "name": "Jan Karres",
-    "email": "mail@jankarres.com",
-    "url": "https://jankarres.com/"
-  },
-  "contributors": [{
-    "name": "Elias Porcio",
-    "email": "elias.porcio@gmail.com"
-  }],
-  "keywords": [
-    "mp3",
-    "wav",
-    "LAME",
-    "MPEG-1",
-    "MPEG-2",
-    "encoder",
-    "audio"
-  ],
-  "license": "MIT",
-  "engines": {
-    "node": ">=8.11"
-  },
-  "scripts": {
-    "build": "tsc",
-    "test": "./node_modules/.bin/mocha"
-  },
-  "files": [
-    "lib/build/",
-    "temp/",
-    "index.js",
-    "index.d.ts",
-    "install.sh",
-    "README.md"
-  ],
-  "dependencies": {},
-  "devDependencies": {
-    "@types/chai": "^3.4.35",
-    "@types/fs-promise": "^1.0.3",
-    "@types/mocha": "^2.2.40",
-    "@types/node": "^8.10.1",
-    "chai": "^3.5.0",
-    "fs-promise": "^2.0.2",
-    "mocha": "^3.2.0",
-    "ts-node": "^2.1.0",
-    "typescript": "^2.2.1"
-  }
+    "name": "node-lame",
+    "description": "LAME MP3 encoder for Node.js",
+    "version": "1.2.0",
+    "homepage": "https://github.com/jankarres/node-lame",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/jankarres/node-lame.git"
+    },
+    "bugs": {
+        "url": "https://github.com/jankarres/node-lame/issues"
+    },
+    "author": {
+        "name": "Jan Karres",
+        "email": "mail@jankarres.com",
+        "url": "https://jankarres.com/"
+    },
+    "contributors": [
+        {
+            "name": "Elias Porcio",
+            "email": "elias.porcio@gmail.com"
+        }
+    ],
+    "keywords": [
+        "mp3",
+        "wav",
+        "LAME",
+        "MPEG-1",
+        "MPEG-2",
+        "encoder",
+        "audio"
+    ],
+    "license": "MIT",
+    "engines": {
+        "node": ">=10.13"
+    },
+    "scripts": {
+        "build": "tsc",
+        "test": "./node_modules/.bin/mocha"
+    },
+    "files": [
+        "lib/build/",
+        "temp/",
+        "index.js",
+        "index.d.ts",
+        "README.md"
+    ],
+    "dependencies": {},
+    "devDependencies": {
+        "@types/chai": "~4.1.7",
+        "@types/fs-extra": "^5.0.4",
+        "@types/mocha": "~5.2.5",
+        "@types/node": "^10.12.5",
+        "chai": "~4.2.0",
+        "fs-extra": "^7.0.1",
+        "mocha": "~5.2.0",
+        "ts-node": "~7.0.1",
+        "typescript": "~3.1.6"
+    }
 }
