@@ -1,13 +1,13 @@
 /**
  * Status data of lame instance
- * 
+ *
  * @interface LameStatus
  */
 interface LameStatus {
-    "started": boolean;
-    "finished": boolean;
-    "progress": number;
-    "eta": string;
+    started: boolean;
+    finished: boolean;
+    progress: number;
+    eta: string;
 }
 
 /**
@@ -21,18 +21,36 @@ namespace Options {
     export type preset = "medium" | "standard" | "extreme" | "insane";
     export type noasm = "mmx" | "3dnow" | "sse";
     export type quality = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
-    export type bitrate = 8 | 16 | 24 | 32 | 40 | 48 | 56 | 64 | 80 | 96 | 112 | 128 | 144 | 160 | 192 | 224 | 256 | 320;
+    export type bitrate =
+        | 8
+        | 16
+        | 24
+        | 32
+        | 40
+        | 48
+        | 56
+        | 64
+        | 80
+        | 96
+        | 112
+        | 128
+        | 144
+        | 160
+        | 192
+        | 224
+        | 256
+        | 320;
     export type emp = "n" | 5 | "c";
     export type resample = 8 | 11.025 | 12 | 16 | 22.05 | 24 | 32 | 44.1 | 48;
 
     export interface meta {
-        "title"?: string;
-        "artist"?: string;
-        "album"?: string;
-        "year"?: string;
-        "comment"?: string;
-        "track"?: string;
-        "genre"?: string;
+        title?: string;
+        artist?: string;
+        album?: string;
+        year?: string;
+        comment?: string;
+        track?: string;
+        genre?: string;
 
         "add-id3v2"?: boolean;
         "id3v1-only"?: boolean;
@@ -47,52 +65,52 @@ namespace Options {
 }
 
 interface Options {
-    "output": string | "buffer";
-    "raw"?: boolean;
+    output: string | "buffer";
+    raw?: boolean;
     "swap-bytes"?: boolean;
-    "sfreq"?: Options.sfreq;
-    "bitwidth"?: Options.bitwidth;
-    "signed"?: boolean;
-    "unsigned"?: boolean;
+    sfreq?: Options.sfreq;
+    bitwidth?: Options.bitwidth;
+    signed?: boolean;
+    unsigned?: boolean;
     "little-endian"?: boolean;
     "big-endian"?: boolean;
-    "mp2Input"?: boolean;
-    "mp3Input"?: boolean;
-    "mode"?: Options.mode;
+    mp2Input?: boolean;
+    mp3Input?: boolean;
+    mode?: Options.mode;
     "to-mono"?: boolean;
     "channel-different-block-sizes"?: boolean;
-    "freeformat"?: Options.freeformat;
+    freeformat?: Options.freeformat;
     "disable-info-tag"?: boolean;
-    "comp"?: number;
-    "scale"?: number;
+    comp?: number;
+    scale?: number;
     "scale-l"?: number;
     "scale-r"?: number;
     "replaygain-fast"?: boolean;
     "replaygain-accurate"?: boolean;
     "no-replaygain"?: boolean;
     "clip-detect"?: boolean;
-    "preset"?: Options.preset;
-    "noasm"?: Options.noasm;
-    "quality"?: Options.quality;
-    "bitrate"?: Options.bitrate;
+    preset?: Options.preset;
+    noasm?: Options.noasm;
+    quality?: Options.quality;
+    bitrate?: Options.bitrate;
     "force-bitrate"?: boolean;
-    "cbr"?: boolean;
-    "abr"?: number;
-    "vbr"?: boolean;
+    cbr?: boolean;
+    abr?: number;
+    vbr?: boolean;
     "vbr-quality"?: number;
     "ignore-noise-in-sfb21"?: boolean;
-    "emp"?: Options.emp;
+    emp?: Options.emp;
     "mark-as-copyrighted"?: boolean;
     "mark-as-copy"?: boolean;
     "crc-error-protection"?: boolean;
-    "nores"?: boolean;
+    nores?: boolean;
     "strictly-enforce-ISO"?: boolean;
-    "lowpass"?: number;
+    lowpass?: number;
     "lowpass-width"?: number;
-    "highpass"?: number;
+    highpass?: number;
     "highpass-width"?: number;
-    "resample"?: Options.resample;
-    "meta"?: Options.meta;
+    resample?: Options.resample;
+    meta?: Options.meta;
 }
 
 export { LameStatus, Options };
